+++ src/mpfr.rs
@@ -1515,10 +1515,10 @@ pub const VERSION_MAJOR: c_int = 4;
 /// See: [`MPFR_VERSION_MINOR`](https://tspiteri.gitlab.io/gmp-mpfr-sys/mpfr/MPFR-Interface.html#index-MPFR_005fVERSION_005fMINOR)
 pub const VERSION_MINOR: c_int = 0;
 /// See: [`MPFR_VERSION_PATCHLEVEL`](https://tspiteri.gitlab.io/gmp-mpfr-sys/mpfr/MPFR-Interface.html#index-MPFR_005fVERSION_005fPATCHLEVEL)
-pub const VERSION_PATCHLEVEL: c_int = 0;
+pub const VERSION_PATCHLEVEL: c_int = 1;
 /// See: [`MPFR_VERSION_STRING`](https://tspiteri.gitlab.io/gmp-mpfr-sys/mpfr/MPFR-Interface.html#index-MPFR_005fVERSION_005fSTRING)
 pub const VERSION_STRING: *const c_char =
-    b"4.0.0\0" as *const u8 as *const c_char;
+    b"4.0.1\0" as *const u8 as *const c_char;
 /// See: [`MPFR_VERSION_NUM`](https://tspiteri.gitlab.io/gmp-mpfr-sys/mpfr/MPFR-Interface.html#index-MPFR_005fVERSION_005fNUM)
 #[inline]
 pub extern "C" fn VERSION_NUM(
@@ -1764,7 +1764,7 @@ mod tests {
 
     #[test]
     fn check_version() {
-        let num_version = "4.0.0";
+        let num_version = "4.0.1";
         let from_constants = format!(
             "{}.{}.{}",
             mpfr::VERSION_MAJOR,
@@ -1772,7 +1772,7 @@ mod tests {
             mpfr::VERSION_PATCHLEVEL
         );
         assert_eq!(from_constants, num_version);
-        let version = "4.0.0";
+        let version = "4.0.1";
         let from_fn = unsafe { CStr::from_ptr(mpfr::get_version()) };
         assert_eq!(from_fn.to_str().unwrap(), version);
         let from_const_string = unsafe { CStr::from_ptr(mpfr::VERSION_STRING) };