+++ RELEASES.md
@@ -1,3 +1,6 @@
+* Fix the type of the `tab` parameter of `mpfr::sum` to
+  `*const *mut mpfr_t` instead of `*mut *mut mpfr_t`.
+
 Version 1.1.0 (2018-01-12)
 ==========================